card, player, players 패키지 생성 및 정리
