+++ src/main/java/domain/card/Suit.java
@@ -1,4 +1,4 @@
-package src.main.java.domain;
+package src.main.java.domain.card;
 
 public enum Suit {
     HEART("하트"), SPADE("스페이드"), CLUB("클로버"), DIAMOND("다이아몬드");
