+++ src/main/java/domain/player/Dealer.java
@@ -1,4 +1,6 @@
-package src.main.java.domain;
+package src.main.java.domain.player;
+
+import src.main.java.domain.card.Card;
 
 public class Dealer extends Player {
